requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
Django==2.1.7
html5lib==0.9999999
Markdown==3.0.1
numpy==1.16.2
Pillow==5.4.1
protobuf==3.7.0
pytz==2018.9
six==1.12.0
tensorflow==1.2.1
Werkzeug==0.14.1
whitenoise==2.0.6
